pyproject.toml
[tool.poetry]
name = "cli-translate-tool"
version = "0.1.0"
description = ""
authors = ["eycjur <eycjur@gmail.com>"]

[tool.poetry.dependencies]
python = "3.10.*"
typer = "^0.7.0"
requests = "^2.28.2"
pyinstaller = "^5.7.0"
chardet = "^5.1.0"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
